Fixes #51
Links from tables now redirect with the full GRB-SN name when both exist. Previously they redirected with the just the SN name or GRB name depending on the link used. This did not provide enough data to search the source files for XRT data. This is because the source files are named GRBXXXXX-SNXXXXxx. This has now been fixed for the homepage and advanced search tables.