Added TNS Spectra.
Got the TNS spectra for 2016jca, 2017iuk and 2017htp, 2018fip, 2019jrj and 2020bvc. I also stored the original format of the data, updated notion and created readmes. I had to change the original filenames to fit with my naming convention and so I have recorded these original names in the readme as they contained info on the date of obs and the instruments used. This data has been added directly to the files.