Fixes #78. Plotfuncs update to handle seconds-seconds.
This handles the case where the time column exists but the time may have been given as a range. It handles seconds-seconds, hours-hours, minutes-minutes and days-days. It calculates the average between the two times and ads it to the first time. A dtime column is added which gets half the delta between the two times.