Standardise data for 117-standardise-data-for-grb130831a-sn2013fu
Standardised data for grb130831a-sn2013fu. Sanity checked with the webtool visually and looks good bar one large error in the optical data which was checked and in the paper too. One point to note was one of the Xray data sets have no errors but in the paper the errors on flux were said to be 90 percent confidence intervals where it was unclear how to calculate them. fixes #117