Standardised data for grb111228a
standardised data for grb111228a. The source 1 paper plots some of the data weirdly where it subtracts a number from the magnitudes making some data look brighter and more distinct which was misleading. There are also chunks of data that are seen clumped together but that is due to the times being either in hours very close together or days very close together. fixes #110