Standardised data for grb200826a SN 2020scz
Radio1 file had flux and no flux density so is not seen plotted on the webtool. Xray 1 data has the counts and xray2 data has flux density at 10keV and not sure how to convert. This broke the web tool and plotfuncs wouldnt run without  flux not sure whether to get rid of. fixes #127