edits to grb200826A
Changed original formats file and found the conversion for the xray 1 data for counts to flux. The flux density problem in xray 2 still not resolved cannot find a conversion factor as those flux density are at an energy of 10eV only and there is no other info. Upper and lower limits not showing for the radio plots either. fixes #127