updates to grb070419a data
Original files renamed and one of the mag units changed from B to AB.  With the html files optical one belonged to a different grb GRB 070518. The multiband data was just a fit with the average flux density for each band in optical and xray at 3000s. Optical 2 had no mag data just times and instruments so is not usable. Only optical 3 had the correct usable data. I have deleted all the html files that were not useful in this update.