Create findbrokenlinks.py
Created a script to look for broken links in the webtool. Code was authored in part by ChatGPT, and inspired by several online sources.

This code is only intended for use in debugging the webtool by the administrators. It does not modify any of the webtool code or the database or observation data.