Rename readme.md as README.md
GitHub seems to want it named this way, and creates a README.md automatically if not.

I suppose this happened because I have edited this file on GitHub before (it doesn't seem to be an issue with the procedurally generated readme.md files for each event for instance). In the past I had realised we had two and deleted the README.md file, but obviously GitHub just recreates it, so renaming should fix that!

This problem was flagged by @joanalnu, so thank you to them!