1998bw adding errors to early time data (#353)
* 1998bw adding errors to early time data

I ran into some trouble with fitting 1998bw as only data over 30 days is in the Optical master txt file. The open supernova catalogue data has the original early time data from Galama et al 1998 but has no errors on the mags. I went to the paper and its stated that there is an error of 0.05 on all VBRI filters and an error of 0.1 on the U. Ive added these errors to all entrys from the 1998 galama paper in the open photometry csv. The read me has been updated to detail this.

* Update readme.yml

* Updated note

Mentioned that it's data from OpenSN

* Updated notes.

---------

Co-authored-by: Gabriel Finneran <gabriel.finneran@ucdconnect.ie>